requirement.txt
customtkinter==5.2.2
pillow==11.3.0
pyinstaller==6.16.0
